+++ src/config/mod.rs
@@ -1,10 +1,12 @@
 pub mod args;
 mod config;
 mod durations;
+mod hooks;
 #[cfg(any(feature = "tcp-client", feature = "tcp-binder"))]
 mod tcp;
 
 pub use config::*;
 pub use durations::*;
+pub use hooks::*;
 #[cfg(any(feature = "tcp-client", feature = "tcp-binder"))]
 pub use tcp::*;