+++ src/config/tcp.rs
@@ -1,12 +1,14 @@
 use serde::{Deserialize, Serialize};
 
-use super::DurationsConfig;
+use super::{DurationsConfig, HooksConfig};
 
 #[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct TcpConfig {
     #[serde(flatten)]
     pub durations: DurationsConfig,
+    #[serde(flatten)]
+    pub hooks: HooksConfig,
     pub host: String,
     pub port: u16,
 }