+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2023-04-14
+
+### Added
+
+- Added hooks support (check
+  https://docs.rs/comodoro/0.0.3/comodoro/config/struct.HooksConfig.html
+  for the list of available hooks).
+
+### Changed
+
+- Improve the way the timer is displayed via the `get` command.
+
 ## [0.0.2] - 2023-04-10
 
 ### Changed